Changed way of checking web password.
Before, I had a dialog box asking for the web password when writing to an ODB value. Unfortunately, the dialog box is produced by the browser and there is no way to hide the password when it is entered. I changed the system now such that at the first write access we get the old method of asking for the password. If it is correctly entered, we will get on the next access the new inline edit method again.